refactor(server): download controller
