+++ server/src/domain/asset/asset.service.ts
@@ -8,7 +8,7 @@ import sanitize from 'sanitize-filename';
 import { AccessCore, Permission } from '../access';
 import { AuthDto } from '../auth';
 import { mimeTypes } from '../domain.constant';
-import { CacheControl, HumanReadableSize, ImmichFileResponse, usePagination } from '../domain.util';
+import { usePagination } from '../domain.util';
 import { IAssetDeletionJob, ISidecarWriteJob, JOBS_ASSET_PAGINATION_SIZE, JobName } from '../job';
 import {
   ClientEvent,
@@ -20,7 +20,6 @@ import {
   IStorageRepository,
   ISystemConfigRepository,
   IUserRepository,
-  ImmichReadStream,
   JobItem,
   TimeBucketOptions,
 } from '../repositories';
@@ -29,15 +28,11 @@ import { SystemConfigCore } from '../system-config';
 import {
   AssetBulkDeleteDto,
   AssetBulkUpdateDto,
-  AssetIdsDto,
   AssetJobName,
   AssetJobsDto,
   AssetOrder,
   AssetSearchDto,
   AssetStatsDto,
-  DownloadArchiveInfo,
-  DownloadInfoDto,
-  DownloadResponseDto,
   MapMarkerDto,
   MemoryLaneDto,
   TimeBucketAssetDto,
@@ -277,111 +272,6 @@ export class AssetService {
     }
 
     return { ...options, userIds };
-  }
-  async downloadFile(auth: AuthDto, id: string): Promise<ImmichFileResponse> {
-    await this.access.requirePermission(auth, Permission.ASSET_DOWNLOAD, id);
-
-    const [asset] = await this.assetRepository.getByIds([id]);
-    if (!asset) {
-      throw new BadRequestException('Asset not found');
-    }
-
-    if (asset.isOffline) {
-      throw new BadRequestException('Asset is offline');
-    }
-
-    return new ImmichFileResponse({
-      path: asset.originalPath,
-      contentType: mimeTypes.lookup(asset.originalPath),
-      cacheControl: CacheControl.NONE,
-    });
-  }
-
-  async getDownloadInfo(auth: AuthDto, dto: DownloadInfoDto): Promise<DownloadResponseDto> {
-    const targetSize = dto.archiveSize || HumanReadableSize.GiB * 4;
-    const archives: DownloadArchiveInfo[] = [];
-    let archive: DownloadArchiveInfo = { size: 0, assetIds: [] };
-
-    const assetPagination = await this.getDownloadAssets(auth, dto);
-    for await (const assets of assetPagination) {
-      // motion part of live photos
-      const motionIds = assets.map((asset) => asset.livePhotoVideoId).filter<string>((id): id is string => !!id);
-      if (motionIds.length > 0) {
-        assets.push(...(await this.assetRepository.getByIds(motionIds)));
-      }
-
-      for (const asset of assets) {
-        archive.size += Number(asset.exifInfo?.fileSizeInByte || 0);
-        archive.assetIds.push(asset.id);
-
-        if (archive.size > targetSize) {
-          archives.push(archive);
-          archive = { size: 0, assetIds: [] };
-        }
-      }
-
-      if (archive.assetIds.length > 0) {
-        archives.push(archive);
-      }
-    }
-
-    return {
-      totalSize: archives.reduce((total, item) => (total += item.size), 0),
-      archives,
-    };
-  }
-
-  async downloadArchive(auth: AuthDto, dto: AssetIdsDto): Promise<ImmichReadStream> {
-    await this.access.requirePermission(auth, Permission.ASSET_DOWNLOAD, dto.assetIds);
-
-    const zip = this.storageRepository.createZipStream();
-    const assets = await this.assetRepository.getByIds(dto.assetIds);
-    const paths: Record<string, number> = {};
-
-    for (const { originalPath, originalFileName } of assets) {
-      const ext = extname(originalPath);
-      let filename = `${originalFileName}${ext}`;
-      const count = paths[filename] || 0;
-      paths[filename] = count + 1;
-      if (count !== 0) {
-        filename = `${originalFileName}+${count}${ext}`;
-      }
-
-      zip.addFile(originalPath, filename);
-    }
-
-    void zip.finalize();
-
-    return { stream: zip.stream };
-  }
-
-  private async getDownloadAssets(auth: AuthDto, dto: DownloadInfoDto): Promise<AsyncGenerator<AssetEntity[]>> {
-    const PAGINATION_SIZE = 2500;
-
-    if (dto.assetIds) {
-      const assetIds = dto.assetIds;
-      await this.access.requirePermission(auth, Permission.ASSET_DOWNLOAD, assetIds);
-      const assets = await this.assetRepository.getByIds(assetIds);
-      return (async function* () {
-        yield assets;
-      })();
-    }
-
-    if (dto.albumId) {
-      const albumId = dto.albumId;
-      await this.access.requirePermission(auth, Permission.ALBUM_DOWNLOAD, albumId);
-      return usePagination(PAGINATION_SIZE, (pagination) => this.assetRepository.getByAlbumId(pagination, albumId));
-    }
-
-    if (dto.userId) {
-      const userId = dto.userId;
-      await this.access.requirePermission(auth, Permission.TIMELINE_DOWNLOAD, userId);
-      return usePagination(PAGINATION_SIZE, (pagination) =>
-        this.assetRepository.getByUserId(pagination, userId, { isVisible: true }),
-      );
-    }
-
-    throw new BadRequestException('assetIds, albumId, or userId is required');
   }
 
   async getStatistics(auth: AuthDto, dto: AssetStatsDto) {
