+++ server/src/domain/asset/dto/index.ts
@@ -2,7 +2,6 @@ export * from './asset-ids.dto';
 export * from './asset-stack.dto';
 export * from './asset-statistics.dto';
 export * from './asset.dto';
-export * from './download.dto';
 export * from './map-marker.dto';
 export * from './memory-lane.dto';
 export * from './time-bucket.dto';
