+++ server/src/domain/download/index.ts
@@ -1,0 +1,2 @@
+export * from './download.dto';
+export * from './download.service';
