+++ server/src/immich/app.module.ts
@@ -19,6 +19,7 @@ import {
   AssetsController,
   AuditController,
   AuthController,
+  DownloadController,
   FaceController,
   JobController,
   LibraryController,
@@ -52,6 +53,7 @@ import { ErrorInterceptor, FileUploadInterceptor } from './interceptors';
     APIKeyController,
     AuditController,
     AuthController,
+    DownloadController,
     FaceController,
     JobController,
     LibraryController,
