+++ server/src/immich/controllers/asset.controller.ts
@@ -13,6 +13,7 @@ import {
   DeviceIdDto,
   DownloadInfoDto,
   DownloadResponseDto,
+  DownloadService,
   MapMarkerDto,
   MapMarkerResponseDto,
   MemoryLaneDto,
@@ -65,7 +66,10 @@ export class AssetsController {
 @Authenticated()
 @UseValidation()
 export class AssetController {
-  constructor(private service: AssetService) {}
+  constructor(
+    private service: AssetService,
+    private downloadService: DownloadService,
+  ) {}
 
   @Get('map-marker')
   getMapMarkers(@Auth() auth: AuthDto, @Query() options: MapMarkerDto): Promise<MapMarkerResponseDto[]> {
@@ -82,31 +86,40 @@ export class AssetController {
     return this.service.getRandom(auth, dto.count ?? 1);
   }
 
+  /**
+   * @deprecated use `/download/info`
+   */
   @SharedLinkRoute()
   @Post('download/info')
-  getDownloadInfo(@Auth() auth: AuthDto, @Body() dto: DownloadInfoDto): Promise<DownloadResponseDto> {
-    return this.service.getDownloadInfo(auth, dto);
+  getDownloadInfoOld(@Auth() auth: AuthDto, @Body() dto: DownloadInfoDto): Promise<DownloadResponseDto> {
+    return this.downloadService.getDownloadInfo(auth, dto);
   }
 
+  /**
+   * @deprecated use `/download/archive`
+   */
   @SharedLinkRoute()
   @Post('download/archive')
   @HttpCode(HttpStatus.OK)
   @FileResponse()
-  downloadArchive(@Auth() auth: AuthDto, @Body() dto: AssetIdsDto): Promise<StreamableFile> {
-    return this.service.downloadArchive(auth, dto).then(asStreamableFile);
+  downloadArchiveOld(@Auth() auth: AuthDto, @Body() dto: AssetIdsDto): Promise<StreamableFile> {
+    return this.downloadService.downloadArchive(auth, dto).then(asStreamableFile);
   }
 
+  /**
+   * @deprecated use `/download/:id`
+   */
   @SharedLinkRoute()
   @Post('download/:id')
   @HttpCode(HttpStatus.OK)
   @FileResponse()
-  async downloadFile(
+  async downloadFileOld(
     @Res() res: Response,
     @Next() next: NextFunction,
     @Auth() auth: AuthDto,
     @Param() { id }: UUIDParamDto,
   ) {
-    await sendFile(res, next, () => this.service.downloadFile(auth, id));
+    await sendFile(res, next, () => this.downloadService.downloadFile(auth, id));
   }
 
   /**
